+++ Software/MATLAB/app/system/+matlab/+compiler/+mlspark/@DataFrameReader/load.m
@@ -1,20 +1,19 @@
 function dataSet = load(obj, varargin)
-% LOAD Method to load the input in as a Dataframe / Dataset
-% The input datasource in as a dataframe / dataset.
-% 
-% For example:
-% 
-%     myDataSet = spark...
-%         .read.format('csv')...
-%         .option('header','true')...
-%         .option('inferSchema','true')...
-%         .load(inputLocation);
+    % LOAD Method to load the input in as a Dataframe / Dataset
+    % The input datasource in as a dataframe / dataset.
+    %
+    % For example:
+    %
+    %     myDataSet = spark...
+    %         .read.format('csv')...
+    %         .option('header','true')...
+    %         .option('inferSchema','true')...
+    %         .load(inputLocation);
 
 
-%                 Copyright 2019 MathWorks, Inc.
-%                 $Id$
+    % Copyright 2019-2022 MathWorks, Inc.
 
-jDataset = obj.dataFrameReader.load(varargin{:});
-dataSet = matlab.compiler.mlspark.Dataset(jDataset);
+    jDataset = obj.dataFrameReader.load(varargin{:});
+    dataSet = matlab.compiler.mlspark.Dataset(jDataset);
 
 end %function